pyproject.toml
[project]
name = "data-agent-proj-extended"
version = "0.1.0"
description = "Extended data analysis tool with GUI"
requires-python = ">=3.10"

dependencies = [
    "langchain>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-openai>=0.3",
    "langgraph>=0.2",
    "pandas>=2.0",
    "pyarrow>=14.0",
    "python-dotenv>=1.0",
    "scikit-learn>=1.3",
    "streamlit>=1.30",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
pythonpath = ["."]
markers = [
    "integration: marks tests as integration tests",
]